requirements.txt
Flask==2.3.3
Flask-PyMongo==2.3.0
Flask-CORS==4.0.0
Flask-Caching==2.1.0
python-dotenv==1.0.0
pymongo==4.6.0
reportlab==4.0.7
requests
gunicorn==23.0.0
